requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
gspread>=6.0.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.0
